# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ToDoApp
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /tasks:
    # our controller name
    x-swagger-router-controller: task
    get:
      operationId: getAllTasks
      description: get the Tasks list
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTaskListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: save
      description: add a new task to the list
      # tasks info to be stored
      parameters:
        - name: taskDescription
          description: Task Description
          in: body
          required: true
          schema:
            $ref: "#/definitions/Task"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /tasks/{id}:
    # our controller name
    x-swagger-router-controller: task
    get:
      operationId: getOneTask
      description: get a task
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTaskResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a task
      # define the parameters
      parameters:
        - name: id
          description: Task id
          type: string
          in: path
          required: true
        - name: taskDescription
          description: Task Description
          in: body
          required: true
          schema:
            $ref: "#/definitions/Task"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delTask
      description: delete a task
      # define the parameters
      parameters:
        - name: id
          description: Task id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  # GET /tasks successful response
  GetTaskListResponse:
    required:
      - tasks
    properties:
      # The array of tasks
      tasks:
        type: array
        items: 
          type: object
          properties:
            id:
              type: string
            taskDescription:
              type: string
            
  GetTaskResponse:
    required:
      - id
      - taskDescription
    properties:
      id:
        type: string
      taskDescription: 
        type: string
  Task:
    type: object
    properties:
      taskDescription:
        type: string
        description: task object name
    required:
      - taskDescription
     
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string